lmsapi/requirements.txt
Django==3.0.7
celery==4.4.1
redis==3.4.1
flower==0.9.3
pytest==5.4.1
pytest-django==3.8.0
djangorestframework
django-cors-headers
django-filter
django-url-filter
mysqlclient
netaddr
pymysql
django-celery-results
gunicorn==20.0.4
